uncomment share button and restore general product details component
